Blitting works. Rasterizer can now draw pixels per given RGB

+++ src/Rasterizer.hpp
@@ -18,12 +18,12 @@ namespace x3d {
     Rasterizer(ScreenInfo *screen_info, unsigned short screen_width, unsigned short screen_height) 
       : sinfo(screen_info), width(screen_width), height(screen_height) { }
     virtual ~Rasterizer() { }
-    void drawPoint(int x, int y, unsigned long color);
-    void drawLine(int x0, int y0, int x1, int y1, unsigned long color);
+    void drawPoint(int x, int y, unsigned int color);
+    void drawLine(int x0, int y0, int x1, int y1, unsigned int color);
 
   protected:
     unsigned char* addressOfPoint(int x, int y);
-    void drawPointAtAddress(unsigned char** p, unsigned long color);
+    void drawPointAtAddress(unsigned char** p, unsigned int color);
   };
 }
 #endif
